Move bouncing ball animation to SoccerBallAnimation component
